move set limit to right in components
