+++ client/src/actions/documentActions/viewDocument.js
