+++ client/src/components/document/index.js
@@ -2,12 +2,12 @@ import CreateDocument from './CreateDocument.jsx';
 import UserDashBoard from './DashBoard.jsx';
 import EditDocument from './EditDocument.jsx';
 import MyDocuments from './MyDocuments.jsx';
-import ViewOneDocument from './ViewOneDocument.jsx';
+import ViewDocument from './ViewDocument.jsx';
 
 export {
   CreateDocument,
   UserDashBoard,
   EditDocument,
   MyDocuments,
-  ViewOneDocument
+  ViewDocument
 };
