+++ client/src/components/user/ViewAllUsers.jsx
@@ -135,7 +135,7 @@ class ViewAllUsers extends React.Component {
                 </Link>
               </div>
             </div>
-            <div className="col s2">
+            <div className="col s2 offset-s10">
               <label htmlFor="limit">Set limit</label>
               <select
                 name="limit"
@@ -161,8 +161,8 @@ class ViewAllUsers extends React.Component {
               deleteUser={this.props.deleteUser}
             />
           </div>
-          <div className="col s12">
-            <center>
+          <div className="col s10 offset-s1">
+            <div className="right">
               {
                 ((this.props.pageCount) ?
                   <Pagination
@@ -175,7 +175,7 @@ class ViewAllUsers extends React.Component {
                     }
                 } /> : '')
               }
-            </center>
+            </div>
           </div>
         </div>
       </div>
