+++ client/src/stylesheet/styles.scss
@@ -370,4 +370,7 @@ header, main, footer, .dashboardContainer {
   white-space: nowrap;
   z-index: 1;
 }
+.red-main {
+  background-color: #F44336 !important;
+}
